Rename ExplosionPrimeEvent - Fixes BUKKIT-2610, BUKKIT-3885
Unblocks BUKKIT-286, BUKKIT-770, BUKKIT-2231.
